Split large editor and rendering files
